requirements.txt
aiogram>=3.3.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
apscheduler>=3.10.0
python-dotenv>=1.0.0